+++ packages/istanbul-widget/types.d.ts
@@ -1,0 +1,9 @@
+import { type IstanbulWidget } from './src/istanbul-widget'
+
+declare global {
+  interface Window {
+    __istanbul_widget__: IstanbulWidget | undefined
+  }
+}
+
+export {}